tests: headings unit testing port
The following ports over the original headings tests from the legacy
container into the unit testing root.
