+++ tests/unit-tests/datasets/common/headings.rst
@@ -1,7 +1,7 @@
 :orphan:
 
 header 1
---------
+========
 
 header 2
-========
+--------
