--- tests/unit-tests/legacy/common/dataset-headings/index.rst
@@ -1,2 +1,0 @@
-base
-----
