--- tests/unit-tests/legacy/common/expected/headings-default.conf
@@ -1,1 +1,0 @@
-<h2>header 2</h2>
