--- tests/unit-tests/legacy/common/expected/headings-with-title.conf
@@ -1,2 +1,0 @@
-<h1>header 1</h1>
-<h2>header 2</h2>
